requirements.txt
openai
pydantic
python-dotenv
aiogram
playwright
sqlalchemy
alembic
aiohttp